Test for async functionality
