+++ tests/client/browse.py
@@ -1,4 +1,4 @@
-from ._cred import TestCaseWithCredentials
+from tests._cred import TestCaseWithCredentials
 from ._resources import artist_ids, category_id, genres, track_id
 
 from tekore.client.api import SpotifyBrowse
