+++ tests/client/library.py
@@ -1,5 +1,5 @@
 from unittest import SkipTest
-from ._cred import TestCaseWithUserCredentials
+from tests._cred import TestCaseWithUserCredentials
 from ._resources import album_ids, track_ids
 
 from tekore.client.api import SpotifyLibrary
