+++ tests/client/player.py
@@ -1,7 +1,7 @@
 from time import sleep
 from requests import HTTPError
 
-from ._cred import TestCaseWithUserCredentials, skip_or_fail
+from tests._cred import TestCaseWithUserCredentials, skip_or_fail
 from ._resources import track_ids, album_id
 from tekore.client.api import SpotifyPlayer
 from tekore.client import Spotify
